bootenv: rename to bootscope
