+++ bootscope/bootscope_test.go
@@ -17,7 +17,7 @@
  *
  */
 
-package bootenv
+package bootscope
 
 import (
 	"testing"
